pyproject.toml
[tool.poetry]
name = "kurs3"
version = "0.1.0"
description = ""
authors = ["<Grachev Den> <<2717087@mail.ru>>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
